requirements.txt
yfinance
pandas
requests
python-dotenv
fastapi
uvicorn
anthropic
pandas-ta
beautifulsoup4
httpx